Prepare resources (#243)
Move the `ResourceHandler` into this project from `creek-platform` and add a `prepare` method.  This allows this project, and `creek-system-test` to call back into client extensions to have them _prepare_ their internal state to handle requests for the supplied resources.

This removes the need for client extensions to have their own logic to collect the resources they handle from the service and aggregate components.  Instead, this is handled by `creek-service` and `creek-system-test`, which already has this 'group-by-resource-type' logic.

This change also fixes some bugs:
 - Catch linkage errors when loading extensions, so that we get a meaningful error message.
 - Don't throw errors if extensions look up their options by some subtype.